# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HEUREKA Real Estate Analysis SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /geocode:
    get:
      operationId: geocodeAddress
      tags: [analyses]
      summary: Geocode an address
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Geocoding results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeocodeResult"
...
components:
  schemas:
...
    GeocodeResult:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/GeocodeItem"
      required:
        - results
    GeocodeItem:
      type: object
      properties:
        label:
          type: string
        score:
          type: number
        housenumber:
          type: string
        street:
          type: string
        postcode:
          type: string
        city:
          type: string
        lat:
          type: number
        lng:
          type: number
      required:
        - label
        - lat
        - lng
...